feat: CAT-206 LikeCommandService랑 구현체 분리
